Extract chatlib for message processing and command returns
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/wricardo/code-surgeon?shareId=XXXX-XXXX-XXXX-XXXX).
